requirements.txt
numpy>=1.21
scipy>=1.7
librosa>=0.10
google-genai>=0.3.0
python-dotenv>=1.0.0
pedalboard>=0.7
flask>=2.0
torch>=2.0
torchaudio>=2.0
transformers>=4.30
